[backend] refactor engine mapping generation (#13635)
